Castling bug and time tests
Fixed a bug with castling where you could castle if an opposing piece captured the rook.

Also implemented a file to test the time required to find positions after a depth of d plies. The results are accurate but the program is extremely slow (most of the time taken up by deepcopying the board). Will try to find a better solution that doesn't require this.